Tweak behavior of Your Report box
So that it is like that of the ATAG Report Tool.

Displays progress on all evaluation pages, and on the overview page if at least one SC is evaluated.

Displays ‘no report started’ on Overview page if no SC was evaluated yet.

Displays ‘Clear report’ button if at least one SC is evaluated.

Signed-off-by: Hidde de Vries <hidde@hiddedevries.nl>